near_swarm/examples/arbitrage.py
"""
Example Arbitrage Strategy
Demonstrates basic arbitrage strategy implementation
"""

from typing import Dict, Any, List, Optional
from datetime import datetime

from near_swarm.core.agent import NEARAgent, AgentConfig
from near_swarm.core.consensus import Vote
from near_swarm.core.memory_manager import StrategyOutcome, MemoryManager


class ArbitrageStrategy:
    """Example arbitrage strategy implementation"""
    
    def __init__(
        self,
        token_pairs: List[str],
        exchanges: List[str],
        min_profit: float = 0.002,
        max_position: float = 10000,
        gas_threshold: float = 0.001
    ):
        """Initialize arbitrage strategy"""
        self.token_pairs = token_pairs
        self.exchanges = exchanges
        self.min_profit = min_profit
        self.max_position = max_position
        self.gas_threshold = gas_threshold
        self.memory = MemoryManager()
    
    async def find_arbitrage(
        self,
        pair: str
    ) -> Dict[str, Any]:
        """Find arbitrage opportunities for a trading pair"""
        try:
            # Get prices from different exchanges
            prices = {}
            opportunities = []
            
            for exchange in self.exchanges:
                prices[exchange] = await self._get_exchange_price(
                    exchange,
                    pair
                )
            
            # Find opportunities
            for buy_exchange in self.exchanges:
                for sell_exchange in self.exchanges:
                    if buy_exchange == sell_exchange:
                        continue
                    
                    buy_price = prices[buy_exchange]
                    sell_price = prices[sell_exchange]
                    
                    profit_pct = (sell_price - buy_price) / buy_price
                    
                    if profit_pct > self.min_profit:
                        opportunities.append({
                            "buy_exchange": buy_exchange,
                            "sell_exchange": sell_exchange,
                            "buy_price": buy_price,
                            "sell_price": sell_price,
                            "profit_pct": profit_pct
                        })
            
            return {
                "pair": pair,
                "opportunities": opportunities,
                "prices": prices
            }
            
        except Exception as e:
            raise Exception(f"Error finding arbitrage: {str(e)}")
    
    async def _get_exchange_price(
        self,
        exchange: str,
        pair: str
    ) -> float:
        """Get token price from an exchange"""
        # Example implementation - replace with actual API calls
        prices = {
            "ref-finance": {
                "NEAR/USDC": 3.50,
                "NEAR/USDT": 3.51,
                "ETH/NEAR": 800.0
            },
            "jumbo": {
                "NEAR/USDC": 3.52,
                "NEAR/USDT": 3.49,
                "ETH/NEAR": 795.0
            }
        }
        return prices[exchange][pair]
    
    async def collect_agent_votes(
        self,
        opportunity: Dict[str, Any]
    ) -> List[Vote]:
        """Collect votes from specialized agents"""
        # Example implementation with mock agents
        agents = [
            {
                "id": "price_validator",
                "confidence": 0.9,
                "decision": True,
                "reasoning": "Price difference exceeds minimum threshold"
            },
            {
                "id": "liquidity_checker",
                "confidence": 0.8,
                "decision": self._evaluate_liquidity(opportunity),
                "reasoning": "Sufficient liquidity available"
            },
            {
                "id": "risk_assessor",
                "confidence": 0.85,
                "decision": True,
                "reasoning": "Risk parameters within acceptable range"
            }
        ]
        
        return [
            Vote(
                agent_id=agent["id"],
                decision=agent["decision"],
                confidence=agent["confidence"],
                reasoning=agent["reasoning"]
            )
            for agent in agents
        ]
    
    def _evaluate_liquidity(
        self,
        opportunity: Dict[str, Any]
    ) -> bool:
        """Evaluate if liquidity is sufficient"""
        # Example implementation - replace with actual liquidity check
        try:
            # Check if exchanges are valid
            if opportunity["opportunity"]["buy_exchange"] not in self.exchanges or \
               opportunity["opportunity"]["sell_exchange"] not in self.exchanges:
                return False
            
            # In a real implementation, check actual liquidity
            return True
            
        except Exception:
            return False
    
    def _calculate_position_size(
        self,
        opportunity: Dict[str, Any],
        gas_cost: float
    ) -> float:
        """Calculate optimal position size"""
        # Start with maximum position
        position = self.max_position
        
        # Scale based on profit potential
        position *= min(
            opportunity["profit_pct"] / self.min_profit,
            1.0
        )
        
        # Adjust for gas costs
        gas_adjusted_profit = opportunity["profit_pct"] - (gas_cost / position)
        if gas_adjusted_profit <= 0:
            return 0
        
        # Apply safety factor
        position *= 0.95  # 5% safety margin
        
        # Further reduce position for high gas costs
        gas_ratio = gas_cost / (position * opportunity["profit_pct"])
        if gas_ratio > self.gas_threshold:
            position *= (self.gas_threshold / gas_ratio)
        
        return min(position, self.max_position)
    
    async def _estimate_gas_costs(
        self,
        pair: str,
        opportunity: Dict[str, Any]
    ) -> float:
        """Estimate gas costs for trades"""
        # Example implementation - replace with actual estimation
        base_cost = 0.1  # NEAR tokens
        
        # Add exchange-specific costs
        exchange_costs = {
            "ref-finance": 0.05,
            "jumbo": 0.06
        }
        
        total_cost = base_cost
        total_cost += exchange_costs[opportunity["buy_exchange"]]
        total_cost += exchange_costs[opportunity["sell_exchange"]]
        
        return total_cost
    
    def _calculate_actual_profit(
        self,
        buy_result: Dict[str, Any],
        sell_result: Dict[str, Any],
        gas_cost: float
    ) -> float:
        """Calculate actual profit from trades"""
        if buy_result["status"] != "success" or sell_result["status"] != "success":
            return 0.0
        
        # Calculate gross profit
        buy_cost = buy_result["amount"] * buy_result["price"]
        sell_revenue = sell_result["amount"] * sell_result["price"]
        gross_profit = sell_revenue - buy_cost
        
        # Subtract gas costs
        net_profit = gross_profit - gas_cost
        
        return max(0.0, net_profit) 